+++ changes.md
@@ -1,5 +1,21 @@
 # Changes
 
+## V4.1.0
+
+### Added
+
+* Lab1: Added `urgent_care_sim.py` a full version of the model built in lab1 for students to browse and use.
+* Lab1: Added `urgent_care_sim.ipynb` a notebook to demo use of `urgent_care_sim.py`
+
+### Changed
+
+* Lab1: Added call centre image to exercise 9.
+
+### Fixed
+
+* Lab 1: fix of minor typos in markdown for solutions and students copy
+* Lab 1: fix of `Patient` and `UrgentCareCallCentre` classes to correctly reference member attribute `self.env`
+
 ## V4.0.0
 
 ## Added